프로젝트 CRUD 수정

+++ ZDrive/Controllers/ProjectController.cs
@@ -33,6 +33,10 @@ public class ProjectController : ControllerBase
             Description = project.Description,
             StartDate = project.StartDate,
             EndDate = project.EndDate,
+            Status = project.Status,
+            Genre = project.Genre,
+            Engine = project.Engine,
+            FileSrc = project.FileSrc,
             UserId = int.Parse(sid)
         };
 
@@ -49,7 +53,9 @@ public class ProjectController : ControllerBase
         if (_project == null)
             return Results.NotFound();
 
-        // Fix me: 프로젝트 Owner만 프로젝트 삭제 편집 가능하게 해야함
+        var sid = User.FindFirstValue(ClaimTypes.Sid);
+        if (sid == null) return Results.Unauthorized();
+        if (sid != _project.UserId.ToString()) return Results.Forbid();
 
         _context.Projects.Remove(_project);
         await _context.SaveChangesAsync();
@@ -59,7 +65,8 @@ public class ProjectController : ControllerBase
     [HttpGet("{id}")]
     public async Task<IResult> Read(int index)
     {
-        var project = await _context.Projects.FindAsync(index);
+        var project = await _context.Projects.Include(p => p.Members)
+            .FirstOrDefaultAsync(p => p.Id == index);
         return project == null ? Results.NotFound() : Results.Ok(project);
     }
 
@@ -78,21 +85,26 @@ public class ProjectController : ControllerBase
 
         if (_project is null) return Results.NotFound();
 
-        // Fix me: 프로젝트 Owner만 프로젝트 삭제 편집 가능하게 해야함
+        var sid = User.FindFirstValue(ClaimTypes.Sid);
+        if (sid == null) return Results.Unauthorized();
+        if (sid != _project.UserId.ToString()) return Results.Forbid();
 
         _project.Name = project.Name;
         _project.Description = project.Description;
         _project.StartDate = project.StartDate;
         _project.EndDate = project.EndDate;
+        _project.Status = project.Status;
+        _project.Genre = project.Genre;
+        _project.Engine = project.Engine;
+        _project.FileSrc = project.FileSrc;
 
         await _context.SaveChangesAsync();
-
         return Results.Created($"/calendar/{_project.Id}", _project);
     }
 
     [Route("member")]
     [HttpPost("{id}")]
-    public async Task<IResult> AddMembers(int id, Project project)
+    public async Task<IResult> AddMembers(int id, Member[] members)
     {
         var _project = await _context.Projects
             .Include(e => e.Members).FirstOrDefaultAsync(e => e.Id == id);
@@ -102,7 +114,7 @@ public class ProjectController : ControllerBase
         if (sid == null) return Results.Unauthorized();
         if (sid != _project.UserId.ToString()) return Results.Forbid();
 
-        foreach (var member in project.Members)
+        foreach (var member in members)
         {
             var studentNum = await _context.StudentNums.FindAsync(member.StudentNumber);
             if (studentNum == null)
@@ -133,6 +145,40 @@ public class ProjectController : ControllerBase
         }
 
         await _context.SaveChangesAsync();
-        return Results.Created<Project>($"/project/{id}", _project);
+        return Results.Created<Project>($"/project/member/{id}", _project);
+    }
+
+    [Route("member")]
+    [HttpPut("{id}")]
+    public async Task<IResult> UpdateMember(int id, Member member)
+    {
+        var _member = await _context.Members.FindAsync(id);
+        if (_member == null) return Results.NotFound();
+
+        var sid = User.FindFirstValue(ClaimTypes.Sid);
+        if (sid == null) return Results.Unauthorized();
+        if (sid != _member.ProjectId.ToString()) return Results.Forbid();
+
+        _member.Description = member.Description;
+        _member.Index = member.Index;
+
+        await _context.SaveChangesAsync();
+        return Results.Created($"/project/member/{id}", member);
+    }
+
+    [Route("member")]
+    [HttpDelete("{id}")]
+    public async Task<IResult> DeleteMember(int id)
+    {
+        var _member = await _context.Members.FindAsync(id);
+        if (_member == null) return Results.NotFound();
+
+        var sid = User.FindFirstValue(ClaimTypes.Sid);
+        if (sid == null) return Results.Unauthorized();
+        if (sid != _member.ProjectId.ToString()) return Results.Forbid();
+
+        _context.Members.Remove(_member);
+        await _context.SaveChangesAsync();
+        return Results.Ok(_member);
     }
 }
